+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 * Give Everyone a Random Buff & Debuff:
   * Now gives several stacks if the random buff/debuff is stackable.
 
+* Add Random Item to Monster Inventory:
+  * Added config option for applying the enemy item blacklist, enabled by default
+
 * One Hit KO:
   * Fixed characters having more than 1 hp if they had barrier before the effect activated
   * All characters will now use the fragile death animation while the effect is active